src/requirements.txt
PyYAML>=6.0
asyncpg>=0.28.0
sqlglot>=25.0.0
sqlparse>=0.5.0
